Make the active step of a multi-step form available to the component via the variable called 'formActiveStepData'. Amend the expression of the ':href' attribute. Amend 'item' watch function.
Make the active step of a multi-step form available to the component via the variable called 'formActiveStepData'. This has been done so that the active menu item can be changed in sync with a multi-step form.  Amend the expression of the ':href' attribute so that the default  text '#' is only used for a link when 'item.href' is not defined. Amend 'item' watch function so that the 'this.active' state can be set even if 'this.item.href' is a blank string.