Make the active step of a multi-step form available to the component via the variable called 'formActiveStepData'. Add 'v-show' attribute with an expression to 'sub-item' and 'item' component.
Make the active step of a multi-step form available to the component via the variable called 'formActiveStepData'. This has been done so that the active menu item can be changed in sync with a multi-step form. Add 'v-show' attribute with an expression to 'sub-item' and 'item' component. This has been done so that menu items can be hidden if access to that form step is restricted.