Amend the expression evaluations for 'this.active' state. Add watch function called 'formActiveStepData'.
Amend the expression evaluations for 'this.active' state.  The 'this.item.href' has been amended so that the 'this.active' can be set even if 'this.item.href' is declared as a blank string. Add watch function called 'formActiveStepData'.  This update the 'this.active' state of the menu item related to the current active step of a multi-step form.